requirements.txt
fastapi
uvicorn[standard]
websockets
prometheus-client
requests
aiohttp
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
langgraph
langchain
langchain-openai
openai
chromadb
numpy
pandas
python-dotenv
httpx
jinja2
pytest
pytest-asyncio
black
flake8
mypy